backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
supabase>=2.0.0
langgraph>=0.0.10
langchain>=0.1.0
langchain-openai>=0.0.1
mcp>=0.1.0
python-dotenv>=1.0.0
httpx>=0.24.0
